// web/admin/src/types/api/v1/models/CreatePromotionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ふるマル API - 管理者向け
 * 管理者向けのふるマルAPIです。
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreatePromotionRequest
 */
export interface CreatePromotionRequest {
    /**
     * 
     * @type {string}
     * @memberof CreatePromotionRequest
     */
    code: string;
    /**
     * 
     * @type {string}
     * @memberof CreatePromotionRequest
     */
    description: string;
    /**
     * 
     * @type {number}
     * @memberof CreatePromotionRequest
     */
    discountRate: number;
    /**
     * 割引計算方法
     * @type {number}
     * @memberof CreatePromotionRequest
     */
    discountType: number;
    /**
     * 
     * @type {number}
     * @memberof CreatePromotionRequest
     */
    endAt: number;
    /**
     * 
     * @type {boolean}
     * @memberof CreatePromotionRequest
     */
    _public: boolean;
    /**
     * 
     * @type {number}
     * @memberof CreatePromotionRequest
     */
    startAt: number;
    /**
     * 
     * @type {string}
     * @memberof CreatePromotionRequest
     */
    title: string;
}

/**
 * Check if a given object implements the CreatePromotionRequest interface.
 */
export function instanceOfCreatePromotionRequest(value: object): value is CreatePromotionRequest {
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('discountRate' in value) || value['discountRate'] === undefined) return false;
    if (!('discountType' in value) || value['discountType'] === undefined) return false;
    if (!('endAt' in value) || value['endAt'] === undefined) return false;
    if (!('_public' in value) || value['_public'] === undefined) return false;
    if (!('startAt' in value) || value['startAt'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    return true;
}

export function CreatePromotionRequestFromJSON(json: any): CreatePromotionRequest {
    return CreatePromotionRequestFromJSONTyped(json, false);
}

export function CreatePromotionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreatePromotionRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'code': json['code'],
        'description': json['description'],
        'discountRate': json['discountRate'],
        'discountType': json['discountType'],
        'endAt': json['endAt'],
        '_public': json['public'],
        'startAt': json['startAt'],
        'title': json['title'],
    };
}

export function CreatePromotionRequestToJSON(json: any): CreatePromotionRequest {
    return CreatePromotionRequestToJSONTyped(json, false);
}

export function CreatePromotionRequestToJSONTyped(value?: CreatePromotionRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'code': value['code'],
        'description': value['description'],
        'discountRate': value['discountRate'],
        'discountType': value['discountType'],
        'endAt': value['endAt'],
        'public': value['_public'],
        'startAt': value['startAt'],
        'title': value['title'],
    };
}