// web/admin/src/types/api/v1/models/UpdatePromotionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ふるマル API - 管理者向け
 * 管理者向けのふるマルAPIです。
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdatePromotionRequest
 */
export interface UpdatePromotionRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdatePromotionRequest
     */
    code: string;
    /**
     * 
     * @type {string}
     * @memberof UpdatePromotionRequest
     */
    description: string;
    /**
     * 
     * @type {number}
     * @memberof UpdatePromotionRequest
     */
    discountRate: number;
    /**
     * 割引計算方法
     * @type {number}
     * @memberof UpdatePromotionRequest
     */
    discountType: number;
    /**
     * 
     * @type {number}
     * @memberof UpdatePromotionRequest
     */
    endAt: number;
    /**
     * 
     * @type {boolean}
     * @memberof UpdatePromotionRequest
     */
    _public: boolean;
    /**
     * 
     * @type {number}
     * @memberof UpdatePromotionRequest
     */
    startAt: number;
    /**
     * 
     * @type {string}
     * @memberof UpdatePromotionRequest
     */
    title: string;
}

/**
 * Check if a given object implements the UpdatePromotionRequest interface.
 */
export function instanceOfUpdatePromotionRequest(value: object): value is UpdatePromotionRequest {
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('discountRate' in value) || value['discountRate'] === undefined) return false;
    if (!('discountType' in value) || value['discountType'] === undefined) return false;
    if (!('endAt' in value) || value['endAt'] === undefined) return false;
    if (!('_public' in value) || value['_public'] === undefined) return false;
    if (!('startAt' in value) || value['startAt'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    return true;
}

export function UpdatePromotionRequestFromJSON(json: any): UpdatePromotionRequest {
    return UpdatePromotionRequestFromJSONTyped(json, false);
}

export function UpdatePromotionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdatePromotionRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'code': json['code'],
        'description': json['description'],
        'discountRate': json['discountRate'],
        'discountType': json['discountType'],
        'endAt': json['endAt'],
        '_public': json['public'],
        'startAt': json['startAt'],
        'title': json['title'],
    };
}

export function UpdatePromotionRequestToJSON(json: any): UpdatePromotionRequest {
    return UpdatePromotionRequestToJSONTyped(json, false);
}

export function UpdatePromotionRequestToJSONTyped(value?: UpdatePromotionRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'code': value['code'],
        'description': value['description'],
        'discountRate': value['discountRate'],
        'discountType': value['discountType'],
        'endAt': value['endAt'],
        'public': value['_public'],
        'startAt': value['startAt'],
        'title': value['title'],
    };
}